Add files via upload
The final group project of the course. This java file creates a player for a game of "Dots and Boxes". My partner and I had to create a player that faced off in a course wide tournament of the game by implementing our own game theory strategy. We finished in the top 8 of 30+ teams.